requirements.txt
Flask==0.10.1
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.7.3
logging==0.4.9.6
logstash_formatter==0.5.8
MarkupSafe==0.23
PyYAML==3.11
requests==2.4.1
Werkzeug==0.9.6
wsgiref==0.1.2

